requirements.txt
Flask
pandas
reportlab
python-kasa
aiohttp
cryptography
